% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{addSingleValue}
\alias{addSingleValue}
\title{add a distinct color for a specific value in the legend color map}
\usage{
addSingleValue(ColMap, s1 = 0.05, s2 = 0.001)
}
\arguments{
\item{ColMap}{a named numeric vector,e.g. \code{c('yellow'=0,'red'=8.99,'green4*'=9)}}

\item{s1, s2}{portion of the whole legend; they are used to control the width
of the color band for the specific value; for example
\itemize{
\item \code{c('yellow'=0,'red'=8.95,'green4*'=9)}, the width is \verb{8.95*(1+s2) to 9}
\item \code{c('yellow*'=0,'red'=0.05,'green4'=9)}, the width is \verb{0 to 0.05*(1-s2)}
\item \code{c('yellow'=0,'red*'=1,'green4'=9)}, the width is \verb{1*(1-s1) to 1*(1+s1)}
}}
}
\value{
a list of two numeric vectors
\itemize{
\item colmap: the adjusted color maps
\item breaks: the breaks for \code{colmap}
}
}
\description{
add a distinct color for a specific value in the legend color map
}
