\name{cheatmap-package}
\alias{cheatmap-package}
\docType{package}
\title{
Comparing two heatmaps
}
\description{
This package will compare two heatmaps.
}
\details{
\tabular{ll}{
Package: \tab cheatmap\cr
Type: \tab Package\cr
Version: \tab 1.0.0 \cr
Date: \tab 2011-12-27 \cr
License: \tab What license is it under?\cr
}
The main interface is cheatmap() function
}
\author{
Xiaoyong Sun
Maintainer: johnsunx1<johnsunx1@gmail.com>
}

\keyword{ package }

\examples{
cheatmap(mtcars, mtcars)
}
