% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{cleanList}
\alias{cleanList}
\title{Remove items of zero length from a list}
\usage{
cleanList(x)
}
\arguments{
\item{x}{a list;}
}
\value{
a list
}
\description{
Remove items of zero length from a list
}
\examples{
#none
}
