% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{colMapDistinctRegions}
\alias{colMapDistinctRegions}
\title{generate color mapping with regions of sharp separation}
\usage{
colMapDistinctRegions(colMap)
}
\arguments{
\item{colMap}{a named numerical vector; colors at point of separation should
be separated by '|'.}
}
\value{
a normal color mapping vector;
}
\description{
For example, it converts this input
\code{c('green'=0,'green4|red'=2,'red4'=4)} to
\code{c('green'=0,'green4'=2,'red'=2+10e-6,'red4'=4)}
}
\examples{
colMapDistinctRegions(c('green'=0,'green4|red'=2,'red4'=4))
}
