% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{formatNum}
\alias{formatNum}
\title{format a numeric vector based on its distribution}
\usage{
formatNum(v, less1 = F, rmLeading0 = F)
}
\arguments{
\item{v}{a numeric vector}

\item{less1}{logical, FALSE; should the number of digits after formatting
be one less than the normal number of digits;}

\item{rmLeading0}{logical, TRUE; removing leading 0 for dicimals.}
}
\value{
a character vector
}
\description{
format a numeric vector based on its distribution
}
\examples{
formatNum(1234.5)
}
