% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{Get distinct colors}
\usage{
getDistinctColors(n)
}
\arguments{
\item{n}{The number of distinct colors}
}
\value{
A character vector of distinct colors
}
\description{
Get distinct colors
}
\examples{
getDistinctColors(5)
}
