% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{getOutliersInd}
\alias{getOutliersInd}
\title{Get the indices of outliers in a numeric vector or matrix}
\usage{
getOutliersInd(x)
}
\arguments{
\item{x}{A numeric vector or matrix}
}
\value{
A list of two int vectors
\enumerate{
\item indices of outliers at the low end
\item indices of outliers at the high end
}
}
\description{
Outliers are < Q1-1.5IQR or > Q3+1.5IQR
}
\examples{
getOutliersInd(c(1,100:110,1000))
}
