% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{gridFuns}
\alias{gridFuns}
\title{a wrapper for \code{\link[grid:grid.points]{grid::grid.points()}} and \code{\link[grid:grid.lines]{grid::grid.lines()}}}
\usage{
gridFuns(x, y, fun1)
}
\arguments{
\item{x, y}{the x and y used in fun1}

\item{fun1}{a list containg info of a grid function e.g.
\code{list('grid.points',size = 0.3,pch=1,col='black')}}
}
\description{
It's to process user-supplied functions. See the doc of the parameter \code{rowDraw}
of \code{cHeatmap} function for details
}
\examples{
# none
}
