% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{hasInfinite}
\alias{hasInfinite}
\title{determine if a vector, matrix, or data.frame contains NA, NaN, Inf, -Inf}
\usage{
hasInfinite(x)
}
\arguments{
\item{x}{a vector, matrix, or data.frame}
}
\value{
a logical value
}
\description{
determine if a vector, matrix, or data.frame contains NA, NaN, Inf, -Inf
}
\examples{
k=c(1:3,Inf)
hasInfinite(k)
k=factor(k)
!is.finite(k)
hasInfinite(k)
}
