% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{nRoundDigits}
\alias{nRoundDigits}
\title{Get the number of rounding digits}
\usage{
nRoundDigits(x)
}
\arguments{
\item{x}{A numeric number}
}
\value{
if abs(x) >= 10, returns 0
else if abs(x) >= 1, returns 1
else returns 2
}
\description{
Get the number of rounding digits
}
\examples{
nRoundDigits(2.38)
}
