% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{setAnnoLegend}
\alias{setAnnoLegend}
\title{Set the color mapping, tick, and tick labels of annotation legends.}
\usage{
setAnnoLegend(
  mat1,
  annoDf,
  annoColMapList = NULL,
  annoParaList = list(),
  on = 1:2,
  rmLegendOutliers = T,
  intAsDiscreteCutoff = 6
)
}
\arguments{
\item{annoDf}{data.frame; annotation}

\item{annoColMapList}{list of color mappings for each annotation, \code{NULL}; its
value is passed to \code{col} in \code{ComplexHeatmap::HeatmapAnnotation()}.
\code{names(annoColMapList)} are among \code{colnames(annoDf)}}

\item{annoParaList}{list of annotation parameters except \code{col} in
\code{ComplexHeatmap::HeatmapAnnotation()}, \code{list()};}

\item{on}{1:2; for row or column annotation}

\item{rmLegendOutliers}{logical, \code{TRUE}; passed to \code{\link[=setColMapAndLegend]{setColMapAndLegend()}}}

\item{intAsDiscreteCutoff}{integer, 6; passed to \code{\link[=setColMapAndLegend]{setColMapAndLegend()}}}
}
\value{
a list() of the following items:
\enumerate{
\item \code{ha}: an object of \code{ComplexHeatmap::HeatmapAnnotation()} where the col and
legend tick and labels are reset based on \code{rmLegendOutliers}
\item \code{newDf}: infinite values in numerical columns of \code{annoDf}, if present, are set to \code{NA}
}
}
\description{
Specifically, it does the following:
\enumerate{
\item for non-numerical variables in \code{annoDf},create the color mapping using
colors in \code{getDistinctColors()} because their default colors are not
very distinct.
\item for numerical variables, apply \code{setColMapAndLegend()}
}
}
\examples{
# none
}
