% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{setColMapAndLegend}
\alias{setColMapAndLegend}
\title{Set the color mapping of a numerical vector, and the tick and tick labels of its legend}
\usage{
setColMapAndLegend(
  vec,
  colMap = NULL,
  whiteValue = NA,
  intAsDiscreteCutoff = 6,
  rmLegendOutliers = T,
  isAnnotation = F,
  legendTicks = NULL,
  legendTickLabels = NULL
)
}
\arguments{
\item{vec}{numerical; a vector or matrix}

\item{colMap}{a named numerical vector for the color mapping,\code{NULL};
if \code{colMap} is not provided, the names of \code{colMap} is set as the followings:
*  if \code{isAnnotation}, \code{c('white',x)} where x is a random color from
\code{getDistinctColors}.
*  otherwise, \code{c('green4', 'white', 'red')}.}

\item{whiteValue}{Numeric, \code{NA}; it is the value of the white or middle color
in the legend if \code{colMap} consists of three colors representing, e.g.
up, no change, down, etc.}

\item{intAsDiscreteCutoff}{integer, 6; if \code{mat1} is an integer matrix and has < 6
unique values, the color mapping in the legend will be discrete.}

\item{rmLegendOutliers}{logical, TRUE; If \code{colMap} is not provided,
if \code{FALSE}, the range of legend ticks covers all values of \code{vac}
if \code{TRUE}, the range of legend ticks covers only non-outliers}

\item{isAnnotation}{logical, \code{FALSE}; if \code{TRUE}, the legend is for the annotation
of a heatmap, not for heatmap itself;}

\item{legendTicks}{numerical vector, NULL;}

\item{legendTickLabels}{character vector, NULL;}
}
\value{
a list of following items
\enumerate{
\item \code{colmap}: the completed \code{colMap}
\item \code{legendParas}: \code{list(at = legendTicks,labels = legendTickLabels)}. it may
also contains \code{legend_gp} if \code{vec} is discrete integers; passed to
\code{ComplexHeatmap::HeatmapAnnotation(annotation_legend_param)}
\item \code{newVals}: infinite values in \code{vec}, if present, are set to \code{NA}
\item \code{isDiscrete}: is \code{vec} discrete?
}
}
\description{
The purpose of this function is to auto-handle outliers in the numerical vectors,
because outliers compress the color mapping and make it difficult to distinguish
among the values of non-outliers.Here are the ways to achieve that goal in this
function if outliers are present:
\enumerate{
\item only non-outliers are used in color-value mapping.
\item a '<' or '>' sign is prefixed to the first or last tick labels.
}
}
\examples{
# none
}
