% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuns.R
\name{vecSwitch}
\alias{vecSwitch}
\title{given a named vector, switch its content and name}
\usage{
vecSwitch(vec1)
}
\arguments{
\item{vec1}{a named vector}
}
\value{
a named vector whose \code{names()} equals \code{vec1} and content is \code{names(vec1)}
}
\description{
given a named vector, switch its content and name
}
\examples{
vecSwitch(c('red'='a','pink'='b','yellow'='c'))
}
